+++ chaosgce/__init__.py
@@ -1,4 +1,13 @@
 # -*- coding: utf-8 -*-
+import warnings
 
-# backward compatibility -> import anything from the new package
-from chaosgcp import *
+from logzero import logger
+
+
+DeprecatedChaosGCEMessage = \
+    "The `chaosgce` package is deprecated and will be removed in a future "\
+    "release. Please use the `chaosgcp` package from now on. "
+
+
+warnings.warn(DeprecatedChaosGCEMessage, DeprecationWarning)
+logger.warn(DeprecatedChaosGCEMessage)